+++ tests/Common/Reader/FilterHandler/EqualsHandlerWithReaderTestTrait.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Data\Tests\Common;
+namespace Yiisoft\Data\Tests\Common\Reader\FilterHandler;
 
 use Yiisoft\Data\Reader\Filter\Equals;
